Renaming Get-PnPFolderFile and Get-PnPFolderFolder as discussed
