+++ src/Commands/Files/GetFileInFolder.cs
@@ -1,45 +1,40 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Linq.Expressions;
 using System.Management.Automation;
 using Microsoft.SharePoint.Client;
 using PnP.Framework.Utilities;
 using PnP.PowerShell.Commands.Base.PipeBinds;
+using File = Microsoft.SharePoint.Client.File;
 using Folder = Microsoft.SharePoint.Client.Folder;
 
 namespace PnP.PowerShell.Commands.Files
 {
-    [Cmdlet(VerbsCommon.Get, "PnPFolderFolder", DefaultParameterSetName = ParameterSet_FOLDERSBYPIPE)]
-    [OutputType(typeof(IEnumerable<Folder>))]
-    public class GetFolderFolder : PnPWebRetrievalsCmdlet<Folder>
+    [Cmdlet(VerbsCommon.Get, "PnPFileInFolder", DefaultParameterSetName = ParameterSet_FOLDERSBYPIPE)]
+    [OutputType(typeof(IEnumerable<File>))]
+    public class GetFileInFolder : PnPWebRetrievalsCmdlet<File>
     {
         private const string ParameterSet_FOLDERSBYPIPE = "Folder via pipebind";
         private const string ParameterSet_FOLDERBYURL = "Folder via url";
 
-        [Parameter(Mandatory = false, ValueFromPipeline = true, Position = 0, ParameterSetName = ParameterSet_FOLDERBYURL)]
+        [Parameter(Mandatory = false, Position = 0, ValueFromPipeline = true, ParameterSetName = ParameterSet_FOLDERBYURL)]
         public string FolderSiteRelativeUrl;
 
-        [Parameter(Mandatory = false, ValueFromPipeline = true, Position = 0, ParameterSetName = ParameterSet_FOLDERSBYPIPE)]
+        [Parameter(Mandatory = false, Position = 0, ValueFromPipeline = true, ParameterSetName = ParameterSet_FOLDERSBYPIPE)]
         public FolderPipeBind Identity;
 
         [Parameter(Mandatory = false)]
         public string ItemName = string.Empty;
 
         [Parameter(Mandatory = false)]
-        public SwitchParameter Recursive;
+        public SwitchParameter Recurse;
 
         [Parameter(Mandatory = false)]
-        public SwitchParameter ExcludeSystemFolders;        
+        public SwitchParameter ExcludeSystemFolders;          
 
         protected override void ExecuteCmdlet()
         {
             CurrentWeb.EnsureProperty(w => w.ServerRelativeUrl);
-
-            if(ExcludeSystemFolders.ToBool())
-            {
-                DefaultRetrievalExpressions = new Expression<Func<Folder, object>>[] { f => f.ListItemAllFields };
-            }
 
             var contents = GetContents(FolderSiteRelativeUrl);
 
@@ -51,20 +46,19 @@ namespace PnP.PowerShell.Commands.Files
             WriteObject(contents, true);
         }
 
-        private IEnumerable<Folder> GetContents(string FolderSiteRelativeUrl)
+        private IEnumerable<File> GetContents(string FolderSiteRelativeUrl)
         {
             Folder targetFolder = null;
             if (string.IsNullOrEmpty(FolderSiteRelativeUrl) && ParameterSetName == ParameterSet_FOLDERSBYPIPE && Identity != null)
             {
                 targetFolder = Identity.GetFolder(CurrentWeb);
-                CurrentWeb.EnsureProperty(w => w.ServerRelativeUrl);
             }
             else
             {
                 string serverRelativeUrl = null;
                 if (!string.IsNullOrEmpty(FolderSiteRelativeUrl))
                 {
-                    serverRelativeUrl = UrlUtility.Combine(CurrentWeb.EnsureProperty(w => w.ServerRelativeUrl), FolderSiteRelativeUrl);
+                    serverRelativeUrl = UrlUtility.Combine(CurrentWeb.ServerRelativeUrl, FolderSiteRelativeUrl);
                 }
 
                 if(string.IsNullOrEmpty(FolderSiteRelativeUrl))
@@ -82,20 +76,27 @@ namespace PnP.PowerShell.Commands.Files
                 }
             }
 
+            IEnumerable<File> files = null;
             IEnumerable<Folder> folders = null;
-            if(ExcludeSystemFolders.ToBool())
-            {
-                folders = ClientContext.LoadQuery(targetFolder.Folders.IncludeWithDefaultProperties(f => f.ListItemAllFields)).Where(f => !ExcludeSystemFolders.ToBool() || !f.ListItemAllFields.ServerObjectIsNull.GetValueOrDefault(false)).OrderBy(f => f.Name);
-            }
-            else
-            {
-                folders = ClientContext.LoadQuery(targetFolder.Folders).OrderBy(f => f.Name);
-            }
-            ClientContext.ExecuteQueryRetry();        
 
-            IEnumerable<Folder> folderContent = folders;
+            files = ClientContext.LoadQuery(targetFolder.Files.IncludeWithDefaultProperties(RetrievalExpressions)).OrderBy(f => f.Name);
 
-            if (Recursive && folders.Count() > 0)
+            if (Recurse)
+            {
+                if(ExcludeSystemFolders.ToBool())
+                {
+                    folders = ClientContext.LoadQuery(targetFolder.Folders.IncludeWithDefaultProperties(f => f.ListItemAllFields)).Where(f => !ExcludeSystemFolders.ToBool() || !f.ListItemAllFields.ServerObjectIsNull.GetValueOrDefault(false)).OrderBy(f => f.Name);
+                }
+                else
+                {
+                    folders = ClientContext.LoadQuery(targetFolder.Folders).OrderBy(f => f.Name);
+                }                
+            }
+            ClientContext.ExecuteQueryRetry();
+
+            IEnumerable<File> folderContent = files;
+
+            if (Recurse && folders.Count() > 0)
             {
                 foreach (var folder in folders)
                 {
@@ -104,7 +105,7 @@ namespace PnP.PowerShell.Commands.Files
                     WriteVerbose($"Processing folder {relativeUrl}");
 
                     var subFolderContents = GetContents(relativeUrl);
-                    folderContent = folderContent.Concat<Folder>(subFolderContents);
+                    folderContent = folderContent.Concat(subFolderContents);
                 }
             }
