+++ src/Commands/Files/GetFolderInFolder.cs
@@ -6,22 +6,21 @@ using System.Management.Automation;
 using Microsoft.SharePoint.Client;
 using PnP.Framework.Utilities;
 using PnP.PowerShell.Commands.Base.PipeBinds;
-using File = Microsoft.SharePoint.Client.File;
 using Folder = Microsoft.SharePoint.Client.Folder;
 
 namespace PnP.PowerShell.Commands.Files
 {
-    [Cmdlet(VerbsCommon.Get, "PnPFolderFile", DefaultParameterSetName = ParameterSet_FOLDERSBYPIPE)]
-    [OutputType(typeof(IEnumerable<File>))]
-    public class GetFolderFile : PnPWebRetrievalsCmdlet<File>
+    [Cmdlet(VerbsCommon.Get, "PnPFolderInFolder", DefaultParameterSetName = ParameterSet_FOLDERSBYPIPE)]
+    [OutputType(typeof(IEnumerable<Folder>))]
+    public class GetFolderInFolder : PnPWebRetrievalsCmdlet<Folder>
     {
         private const string ParameterSet_FOLDERSBYPIPE = "Folder via pipebind";
         private const string ParameterSet_FOLDERBYURL = "Folder via url";
 
-        [Parameter(Mandatory = false, Position = 0, ValueFromPipeline = true, ParameterSetName = ParameterSet_FOLDERBYURL)]
+        [Parameter(Mandatory = false, ValueFromPipeline = true, Position = 0, ParameterSetName = ParameterSet_FOLDERBYURL)]
         public string FolderSiteRelativeUrl;
 
-        [Parameter(Mandatory = false, Position = 0, ValueFromPipeline = true, ParameterSetName = ParameterSet_FOLDERSBYPIPE)]
+        [Parameter(Mandatory = false, ValueFromPipeline = true, Position = 0, ParameterSetName = ParameterSet_FOLDERSBYPIPE)]
         public FolderPipeBind Identity;
 
         [Parameter(Mandatory = false)]
@@ -31,11 +30,16 @@ namespace PnP.PowerShell.Commands.Files
         public SwitchParameter Recursive;
 
         [Parameter(Mandatory = false)]
-        public SwitchParameter ExcludeSystemFolders;          
+        public SwitchParameter ExcludeSystemFolders;        
 
         protected override void ExecuteCmdlet()
         {
             CurrentWeb.EnsureProperty(w => w.ServerRelativeUrl);
+
+            if(ExcludeSystemFolders.ToBool())
+            {
+                DefaultRetrievalExpressions = new Expression<Func<Folder, object>>[] { f => f.ListItemAllFields };
+            }
 
             var contents = GetContents(FolderSiteRelativeUrl);
 
@@ -47,19 +51,20 @@ namespace PnP.PowerShell.Commands.Files
             WriteObject(contents, true);
         }
 
-        private IEnumerable<File> GetContents(string FolderSiteRelativeUrl)
+        private IEnumerable<Folder> GetContents(string FolderSiteRelativeUrl)
         {
             Folder targetFolder = null;
             if (string.IsNullOrEmpty(FolderSiteRelativeUrl) && ParameterSetName == ParameterSet_FOLDERSBYPIPE && Identity != null)
             {
                 targetFolder = Identity.GetFolder(CurrentWeb);
+                CurrentWeb.EnsureProperty(w => w.ServerRelativeUrl);
             }
             else
             {
                 string serverRelativeUrl = null;
                 if (!string.IsNullOrEmpty(FolderSiteRelativeUrl))
                 {
-                    serverRelativeUrl = UrlUtility.Combine(CurrentWeb.ServerRelativeUrl, FolderSiteRelativeUrl);
+                    serverRelativeUrl = UrlUtility.Combine(CurrentWeb.EnsureProperty(w => w.ServerRelativeUrl), FolderSiteRelativeUrl);
                 }
 
                 if(string.IsNullOrEmpty(FolderSiteRelativeUrl))
@@ -77,25 +82,18 @@ namespace PnP.PowerShell.Commands.Files
                 }
             }
 
-            IEnumerable<File> files = null;
             IEnumerable<Folder> folders = null;
-
-            files = ClientContext.LoadQuery(targetFolder.Files.IncludeWithDefaultProperties(RetrievalExpressions)).OrderBy(f => f.Name);
-
-            if (Recursive)
+            if(ExcludeSystemFolders.ToBool())
             {
-                if(ExcludeSystemFolders.ToBool())
-                {
-                    folders = ClientContext.LoadQuery(targetFolder.Folders.IncludeWithDefaultProperties(f => f.ListItemAllFields)).Where(f => !ExcludeSystemFolders.ToBool() || !f.ListItemAllFields.ServerObjectIsNull.GetValueOrDefault(false)).OrderBy(f => f.Name);
-                }
-                else
-                {
-                    folders = ClientContext.LoadQuery(targetFolder.Folders).OrderBy(f => f.Name);
-                }                
+                folders = ClientContext.LoadQuery(targetFolder.Folders.IncludeWithDefaultProperties(f => f.ListItemAllFields)).Where(f => !ExcludeSystemFolders.ToBool() || !f.ListItemAllFields.ServerObjectIsNull.GetValueOrDefault(false)).OrderBy(f => f.Name);
             }
-            ClientContext.ExecuteQueryRetry();
+            else
+            {
+                folders = ClientContext.LoadQuery(targetFolder.Folders).OrderBy(f => f.Name);
+            }
+            ClientContext.ExecuteQueryRetry();        
 
-            IEnumerable<File> folderContent = files;
+            IEnumerable<Folder> folderContent = folders;
 
             if (Recursive && folders.Count() > 0)
             {
@@ -106,7 +104,7 @@ namespace PnP.PowerShell.Commands.Files
                     WriteVerbose($"Processing folder {relativeUrl}");
 
                     var subFolderContents = GetContents(relativeUrl);
-                    folderContent = folderContent.Concat<File>(subFolderContents);
+                    folderContent = folderContent.Concat<Folder>(subFolderContents);
                 }
             }
